refactor(tools): split card-tools.ts into focused modules (#66)
Split 1,109-line card-tools.ts into 6 focused modules:
- card-crud-tools.ts: list, get, create, update, delete + metadata
- card-move-tools.ts: move card, get/set size
- card-comment-tools.ts: comment CRUD
- card-subtask-tools.ts: subtask operations
- card-relationship-tools.ts: parents, children, linked cards
- card-bulk-tools.ts: bulk delete/update

Uses delegation pattern - CardToolHandler delegates to module
registration functions while maintaining backward compatibility.
